Add files via upload
Problem
Chef owns a car that can run 
15
15 kilometers using 
1
1 liter of petrol.
He wants to attend a programming camp at DAIICT.

�
X liters of petrol is present in Chef's car. The distance between his house and DAIICT is 
�
Y kilometers.
Determine whether Chef can attend the event and return to his home with the given amount of petrol.

Note: Chef has to return back to home, so the total distance to be covered would be 
2
⋅
�
2⋅Y.

Input Format
The first line of input will contain a single integer 
�
T, denoting the number of test cases.
Each test case consists of two space-separated integers 
�
X and 
�
Y — the amount of petrol in liter and the distance between Chef's house and DAIICT in kilometers, respectively.
Output Format
For each test case, print YES if it is possible for the chef to attend the event and return home, else, print NO.

You may print each character in uppercase or lowercase. For example, Yes, yes, YES, and yES are all considered identical.

Explanation:
Test case 
1
1: Chef needs to cover a distance of 
100
100 kilometers in total. He can cover it with 
10
10 liters of petrol.

Test case 
2
2: Chef needs to cover a distance of 
400
400 kilometers in total but using 
15
15 liters of petrol, he can cover only 
225
225 kilometers.

Test case 
3
3: Chef needs to cover a distance of 
40
40 kilometers in total. He can cover it with 
3
3 liters of petrol.