Difficulty Rating:841
Problem
Read problem statements in Bengali, Mandarin Chinese, Russian, and Vietnamese as well.
Chef has 
�
A units of solid and 
�
B units of liquid. He combines them to create a mixture. What kind of mixture does Chef produce: a solution, a solid, or a liquid?

A mixture is called :

A solution if 
�
>
0
A>0 and 
�
>
0
B>0,

A solid if 
�
=
0
B=0, or

A liquid if 
�
=
0
A=0.

Input Format
The first line contains 
�
T denoting the number of test cases. Then the test cases follow.
Each test case contains two space-separated integers 
�
A and 
�
B on a single line.
Output Format
For each test case, output on a single line the type of mixture Chef produces, whether it is a Solution, Solid, or Liquid. The output is case sensitive.

Explanation:
Test case 
1
1: Chef adds both solid and liquid to the mixture, hence the mixture is a solution.

Test case 
2
2: Chef does not add solid to the mixture, hence the mixture is liquid.

Test case 
3
3: Chef does not add liquid to the mixture, hence the mixture is solid.